雪球实体系统大检查与综合bug修复

+++ src/main/java/com/linngdu664/bsf3lite/entity/snowball/util/ILaunchAdjustment.java
@@ -1,7 +1,7 @@
 package com.linngdu664.bsf3lite.entity.snowball.util;
 
 public interface ILaunchAdjustment {
-    double adjustPunch(double punch);
+    float adjustPunch(float punch);
 
     int adjustWeaknessTicks(int weaknessTicks);
