+++ src/main/java/com/linngdu664/bsf3lite/item/weapon/ImplosionSnowballCannonItem.java
@@ -123,7 +123,7 @@ public class ImplosionSnowballCannonItem extends AbstractBSFWeaponItem {
     }
 
     @Override
-    public ILaunchAdjustment getLaunchAdjustment(double damageDropRate, Item snowball) {
+    public ILaunchAdjustment getLaunchAdjustment(float damageDropRate, Item snowball) {
         return null;
     }
