+++ src/main/java/com/linngdu664/bsf3lite/item/weapon/SnowballShotgunItem.java
@@ -45,11 +45,11 @@ public class SnowballShotgunItem extends AbstractBSFWeaponItem {
     }
 
     @Override
-    public ILaunchAdjustment getLaunchAdjustment(double damageDropRate, Item snowball) {
+    public ILaunchAdjustment getLaunchAdjustment(float damageDropRate, Item snowball) {
         return new ILaunchAdjustment() {
             @Override
-            public double adjustPunch(double punch) {
-                return punch + 1.51;
+            public float adjustPunch(float punch) {
+                return punch + 1.51F;
             }
 
             @Override
