+++ src/main/java/com/linngdu664/bsf3lite/item/weapon/cannon/FreezingSnowballCannonItem.java
@@ -24,11 +24,11 @@ public class FreezingSnowballCannonItem extends SnowballCannonItem {
     }
 
     @Override
-    public ILaunchAdjustment getLaunchAdjustment(double damageDropRate, Item snowball) {
+    public ILaunchAdjustment getLaunchAdjustment(float damageDropRate, Item snowball) {
         return new ILaunchAdjustment() {
             @Override
-            public double adjustPunch(double punch) {
-                return punch + damageDropRate * 1.51;
+            public float adjustPunch(float punch) {
+                return punch + damageDropRate * 1.51F;
             }
 
             @Override
