+++ src/main/java/com/linngdu664/bsf3lite/item/weapon/cannon/PowerfulSnowballCannonItem.java
@@ -24,11 +24,11 @@ public class PowerfulSnowballCannonItem extends SnowballCannonItem {
     }
 
     @Override
-    public ILaunchAdjustment getLaunchAdjustment(double damageDropRate, Item snowball) {
+    public ILaunchAdjustment getLaunchAdjustment(float damageDropRate, Item snowball) {
         return new ILaunchAdjustment() {
             @Override
-            public double adjustPunch(double punch) {
-                return punch + damageDropRate * 2.5;
+            public float adjustPunch(float punch) {
+                return punch + damageDropRate * 2.5F;
             }
 
             @Override
