+++ src/main/java/com/linngdu664/bsf3lite/item/weapon/rifle/FreezingSnowballRifle.java
@@ -14,11 +14,11 @@ public class FreezingSnowballRifle extends SnowballRifleItem {
     }
 
     @Override
-    public ILaunchAdjustment getLaunchAdjustment(double damageDropRate, Item snowball) {
+    public ILaunchAdjustment getLaunchAdjustment(float damageDropRate, Item snowball) {
         return new ILaunchAdjustment() {
             @Override
-            public double adjustPunch(double punch) {
-                return punch + 1.2;
+            public float adjustPunch(float punch) {
+                return punch + 1.2F;
             }
 
             @Override
