+++ src/main/java/com/linngdu664/bsf3lite/item/weapon/rifle/SnowballRifleItem.java
@@ -43,11 +43,11 @@ public class SnowballRifleItem extends AbstractBSFWeaponItem {
     }
 
     @Override
-    public ILaunchAdjustment getLaunchAdjustment(double damageDropRate, Item snowball) {
+    public ILaunchAdjustment getLaunchAdjustment(float damageDropRate, Item snowball) {
         return new ILaunchAdjustment() {
             @Override
-            public double adjustPunch(double punch) {
-                return punch + 1.2;
+            public float adjustPunch(float punch) {
+                return punch + 1.2F;
             }
 
             @Override
